+++ lib/decorators/rmq-controller.decorator.ts
@@ -1,6 +1,6 @@
 export function RMQController(): ClassDecorator {
 	return function (target: any) {
-		console.warn('@RMQController is deprecated. Use just @RMQRoute. You can use it with msgFactory inside if needed');
+		console.warn('@RMQController is deprecated. Use just @RMQRoute. You can use it with msgFactory inside if needed', 'RMQModule');
 		target = class extends (target as { new (...args): any }) {
 			constructor(...args: any) {
 				super(...args);